requirements.txt
flask==3.0.2
langchain-core>=0.1.7
langchain-community>=0.0.13
langchain-openai>=0.0.2
faiss-cpu==1.8.0
openai>=1.6.1
unstructured==0.12.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
langchain-text-splitters>=0.0.1 